Implements potentiometer, servos, and a new command to force the launcher down while going under obstacles

+++ src/org/usfirst/frc/team4915/stronghold/commands/IntakeLauncher/SetSetPointFromSmartDashboardCommand.java
@@ -1,9 +1,8 @@
 package org.usfirst.frc.team4915.stronghold.commands.IntakeLauncher;
 
-import org.usfirst.frc.team4915.stronghold.Robot;
-
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
+import org.usfirst.frc.team4915.stronghold.Robot;
 
 /**
  *
@@ -11,30 +10,26 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 public class SetSetPointFromSmartDashboardCommand extends Command {
 
     public SetSetPointFromSmartDashboardCommand() {
-        // Use requires() here to declare subsystem dependencies
-        // eg. requires(chassis);
+    
     }
 
-    // Called just before this Command runs the first time
     protected void initialize() {
-        Robot.intakeLauncher.setSetPoint((int)SmartDashboard.getNumber("Launcher Set Point: "));
+        Robot.intakeLauncher.setSetPoint(SmartDashboard.getNumber("Launcher Set Point: "));
     }
 
-    // Called repeatedly when this Command is scheduled to run
     protected void execute() {
+    
     }
 
-    // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
         return true;
     }
 
-    // Called once after isFinished returns true
     protected void end() {
+    
     }
 
-    // Called when another command which requires one or more of the same
-    // subsystems is scheduled to run
     protected void interrupted() {
+    
     }
 }
